+++ src/neunet/loader.rs
@@ -3,5 +3,5 @@ use std::io::*;
 use nalgebra::*;
 
 pub trait DataLoader {
-    fn load_data(self, data_path: String, labels_path: String) -> Result<(DMatrix<f64>, DVector<u8>)>;
+    fn load_data(self, data_path: String, labels_path: String) -> Result<(DMatrix<f32>, DVector<u8>)>;
 }
